Tests/LibWeb/Text/input/WebAudio/OfflineAudioContext-AudioWorklet-process.html

<!doctype html>
<script src="../include.js"></script>
<script>
    asyncTest(async done => {
        const context = new OfflineAudioContext(1, 128, 44100);

        await context.audioWorklet.addModule("../../data/AudioWorklet-constant-gain-processor.js");

        const node = new AudioWorkletNode(context, "constant-gain");
        node.parameters.get("gain").value = 0.25;

        node.connect(context.destination);

        const buffer = await context.startRendering();
        const data = buffer.getChannelData(0);

        println("sample0: " + data[0].toFixed(2));
        println("sample1: " + data[1].toFixed(2));
        println("sample2: " + data[2].toFixed(2));
        done();
    });
</script>
