Tests/LibWeb/Text/expected/WebAudio/OfflineAudioContext-AudioWorklet-process.txt

sample0: 0.25
sample1: 0.25
sample2: 0.25
